refactor: Refactor click event handler of app
